arrow-row: Move REE tests from separate file next to implementation
